Refactor Reactions to component
